+++ stac_fastapi/extensions/stac_fastapi/extensions/core/filter/__init__.py
@@ -1,0 +1,6 @@
+"""Filter extension module."""
+
+
+from .filter import FilterExtension
+
+__all__ = ["FilterExtension"]