+++ CHANGES.md
@@ -6,6 +6,9 @@
 
 ### Changed
 
+* Refactor to remove hardcoded search request models. Request models are now dynamically created based on the enabled extensions.
+  ([#213](https://github.com/stac-utils/stac-fastapi/pull/213))  
+
 ### Removed
 
 ### Fixed